Files | Split file tab into components Assumes that ShokoAnime/ShokoServer#1128 goes through
